1、注册功能修改内存溢出的问题 2、API2改版（更新，退出，支付，订单，设置，编辑图片）

+++ app/src/main/java/com/pictureair/photopass/activity/RegisterOrForgetActivity.java
@@ -276,7 +276,7 @@ public class RegisterOrForgetActivity extends BaseActivity implements RegisterOr
     }
 
     @Override
-    public void countDown(int time) {
+    public void countDown(long time) {
         if (time == 0) {
             countDownFinish = true;
             btn_next.setText(R.string.smssdk_send_verification_code);// 再次发送验证码
